migration fixes to workaround foreign key problem in rollback

+++ database/migrations/2024_12_29_132300_add_current_activity_to_users_table.php
@@ -13,7 +13,9 @@ return new class extends Migration
     {
         Schema::table('users', function (Blueprint $table) {
             //keep track of lowest non-optional incomplete activity
-            $table->foreignId('current_activity')->nullable()->constrained('activities')->onDelete('set null');
+            if (!Schema::hasColumn('users', 'current_activity')) {
+                $table->foreignId('current_activity')->nullable()->constrained('activities')->onDelete('set null');
+            }
         });
     }
 
@@ -23,7 +25,7 @@ return new class extends Migration
     public function down(): void
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->dropColumn('current_activity');
+            // $table->dropColumn('current_activity');
         });
     }
 };
